attack: add basics for opening channel and identifying by alias
attack: add balance on attacker side to prevent hitting fee buffer

If we have very little balance on our side of the channel, LND will
stop forwarding non-dust htlcs to us because we have insufficient fees
to pay for the htlcs in event of a fee spike.

This commit splits our balance in half so that we don't need to worry
about this.

attack: open two channels with peer to ensure we can jam protected

When we start to attack the target -> peer channel's general slots,
it's possible that some of our HTLCs will not make it into the protected
slots of the peer to LND2 channel (because the target does not have
sufficient reputation for all of the HTLCs to get endorsed, only some).

We open two channels so that there's always room in the general
resource bucket for all of our jamming payments on the hop to LND2.