attack: add balance on attacker side to prevent hitting fee buffer
If we have very little balance on our side of the channel, LND will
stop forwarding non-dust htlcs to us because we have insufficient fees
to pay for the htlcs in event of a fee spike.

This commit splits our balance in half so that we don't need to worry
about this.