setup: pre-generate 2 weeks of projected data for simulations
When we generate historical data using sim-ln for a given bootstrap
period, we can also project this generation forward so that we have a
record of what the network would have processed if _not_ under attack.
When sim-ln has the ability to run with a fixed seed (waiting on PR
https://github.com/bitcoin-dev-project/sim-ln/pull/178), we'll be able
to run this generation and the actual network under attack with the
same seed. This allows us to directly compare the network with and
without the attack, because the generated traffic will be the same.

This is of course subject to the differences between the simulator and
an actual network - latency is different, so HTLCs will lock in
slightly different orders - but it gives us a good ballpark.