attack: open two channels with peer to ensure we can jam protected
When we start to attack the target -> peer channel's general slots,
it's possible that some of our HTLCs will not make it into the protected
slots of the peer to LND2 channel (because the target does not have
sufficient reputation for all of the HTLCs to get endorsed, only some).

We open two channels so that there's always room in the general
resource bucket for all of our jamming payments on the hop to LND2.